Recover from problems with the server connection
There is much clean-up and testing needed, but this should be a massive
improvement in stability. I would like to investigate further whether
the default Akka pool actors can be used reliably with actors like
RedisConnectionActor that manage resources that are not immediately
ready for use, and where all resources may be unusable for a time.
But I must say that I've lost some hope. Perhaps more of the built-in
classes can be made use of though? We'll have to see.