Use updated Redis type naming
Switched from "reply"-based names to the type names that are now used in
the Redis documentation, prefixed with "R".

Added RValueSerializer for serializing values. This is now used for
sending commands (which are represented as RArray's of RBulkString's on
the wire).

Replaced the commented-out, incomplete parser implementation from
PushParserReplyReconstructor with the new RValueParser. However, the
ParserCombinatorReplyReconstructor is still being used in the
RedisConnectionActor for the time being.