Simplify loopback address code
This new code results in an IPv6 loopback addresses without zone
identifiers.

This should fix the Oracle JDK 8 build, which seems to be broken because
starting with JDK 8, IPv6 addresses are being represented with
non-numeric zone identifiers suffixes such as “%lo”, whereas previously
they were numeric (e.g. “%1”). It seems that Redis 2.8.17 on Linux
doesn’t like this in arguments to the --bind option.