Fix MiMa integration in release.sh
I think I might have broken this by upgrading the MiMa sbt plugin.

This also moves building before release notes preparation, which while
anoying from a timing point of view, is nice in that the MiMa output is
now included in the draft release notes.

Retrieving the previous artefacts for MiMa to test against is now done
before testing, to allow us to fail fast when this fails.